# src/routes/admin-routes.js
import { Router } from 'express';
import crypto from 'node:crypto';
import db, { hashApiKey, id, now, verifyPassword } from '../database.js';
import { authorize, cookieName, tokenHash } from '../middleware/authorization.js';
import { config } from '../config.js';
import { publicPayment } from '../services/payment-service.js';
import { PaymentMethodService } from '../services/payment-method-service.js';
import { deleteMerchant } from '../services/merchant-deletion-service.js';
import { ipAllowlistService } from '../services/ip-allowlist-service.js';
import { requirePermission } from '../middleware/permissions.js';
import { loginThrottle } from '../middleware/login-throttle.js';
import { validateOutboundUrl } from '../services/outbound-url-policy.js';

const allowedStatuses = new Set(['PENDING', 'PAID', 'FAILED', 'EXPIRED']);
const parseCookies = (header = '') => Object.fromEntries(header.split(';').map((x) => x.trim().split('=').map(decodeURIComponent)).filter((x) => x.length === 2));
const safeMerchant = (merchant) => ({ id: merchant.id, name: merchant.name, callback_url: merchant.callback_url, is_active: Boolean(merchant.is_active), created_at: merchant.created_at });
const validCallbackUrl = (value) => { if (!value) return true; try { validateOutboundUrl(value); return true; } catch { return false; } };
const newApiKey = () => `np_live_${crypto.randomBytes(20).toString('hex')}`;
const csvCell = (value) => { const text = String(value ?? ''); const safe = /^[=+\-@]/.test(text) ? `'${text}` : text; return `"${safe.replaceAll('"', '""')}"`; };

const auth = authorize();
async function merchantById(merchantId, res) {
  const merchant = await db.get('SELECT * FROM merchants WHERE id=?',[merchantId]);
  if (!merchant) { res.status(404).json({ error: { message: 'Merchant tidak ditemukan.' } }); return null; }
  return merchant;
}
function paymentFilters(query) {
  const clauses = []; const params = [];
  if (query.status && allowedStatuses.has(query.status)) { clauses.push('p.status=?'); params.push(query.status); }
  if (query.merchant_id) { clauses.push('p.merchant_id=?'); params.push(query.merchant_id); }
  if (query.search?.trim()) { clauses.push('(p.id LIKE ? OR p.merchant_order_id LIKE ? OR p.provider_reference LIKE ?)'); const value = `%${query.search.trim()}%`; params.push(value, value, value); }
  if (query.from) { clauses.push('p.created_at>=?'); params.push(`${query.from}T00:00:00.000Z`); }
  if (query.to) { clauses.push('p.created_at<=?'); params.push(`${query.to}T23:59:59.999Z`); }
  return { where: clauses.length ? `WHERE ${clauses.join(' AND ')}` : '', params };
}
function adminPayment(row) { return { ...publicPayment(row), merchant_id: row.merchant_id, merchant_name: row.merchant_name, provider_transaction_id: row.provider_transaction_id, customer: { name: row.customer_name, email: row.customer_email, phone: row.customer_phone } }; }
async function auditIpRule(req,action,rule){await db.run('INSERT INTO audit_logs (id,actor_id,action,target_type,target_id,ip,user_agent,metadata_json,created_at) VALUES (?,?,?,?,?,?,?,?,?)',[id('aud'),req.admin.user_id,action,'ip_allowlist',rule.id,req.ip,req.get('user-agent'),JSON.stringify({merchant_id:rule.merchant_id,cidr:rule.cidr,label:rule.label,is_active:rule.is_active}),now()]);}

export default function adminRoutes(webhookService, paymentService) {
  const router = Router();
  const paymentMethods = new PaymentMethodService();
  router.post('/login', loginThrottle, async (req, res, next) => { try { const username = req.body?.username?.trim() || req.body?.email?.trim() || ''; const user = await db.get('SELECT * FROM admin_users WHERE (username=? OR email=?) AND is_active=1', [username, username]); if (!user || !verifyPassword(req.body?.password, user.password_hash)){req.recordLoginFailure();return res.status(401).json({ error: { message: 'Username atau password tidak valid.' } });} req.resetLoginThrottle(); await db.run('DELETE FROM admin_sessions WHERE expires_at<=?', [new Date().toISOString()]); const token=crypto.randomBytes(32).toString('base64url'); const expires=new Date(Date.now()+config.sessionTtlMs).toISOString(); await db.run('INSERT INTO admin_sessions (id,user_id,token_hash,expires_at,ip,user_agent,created_at) VALUES (?,?,?,?,?,?,?)',[id('ses'),user.id,tokenHash(token),expires,req.ip,req.get('user-agent'),now()]); await db.run('UPDATE admin_users SET last_login_at=? WHERE id=?',[now(),user.id]); res.setHeader('Set-Cookie', `${cookieName}=${token}; HttpOnly; SameSite=Strict; Path=/; Max-Age=${Math.floor(config.sessionTtlMs/1000)}${config.production?'; Secure':''}`); res.json({ data: { authenticated:true, user:{name:user.name,email:user.email,username:user.username,role:user.role},expires_at:expires } }); } catch (error) { next(error); } });
  router.get('/session', auth, (req, res) => { res.json({ data: { authenticated: true, user: { name: req.admin.name, email: req.admin.email, role: req.admin.role }, expires_at: req.admin.expires_at } }); });
  router.post('/logout', auth, async (req, res, next) => { try { await db.run('DELETE FROM admin_sessions WHERE token_hash=?',[tokenHash(req.adminToken)]); res.setHeader('Set-Cookie', `${cookieName}=; HttpOnly; SameSite=Strict; Path=/; Max-Age=0${config.production?'; Secure':''}`); res.json({ data: { logged_out: true } }); } catch (error) { next(error); } });
  router.get('/overview', auth, requirePermission('overview:read'), async (req, res, next) => { try {
    const parseDate = (value, fallback) => { if (!value) return fallback; const parsed = new Date(`${value}T00:00:00.000Z`); return Number.isNaN(parsed.getTime()) ? null : parsed; };
    const endDefault = new Date(); const startDefault = new Date(endDefault); startDefault.setUTCDate(startDefault.getUTCDate() - 29);
    const start = parseDate(req.query.from, startDefault); const end = parseDate(req.query.to, endDefault);
    if (!start || !end || start > end) return res.status(422).json({ error: { message: 'Rentang tanggal tidak valid.' } });
    const rangeDays = Math.floor((end - start) / 86400000) + 1; if (rangeDays > 366) return res.status(422).json({ error: { message: 'Rentang statistik maksimal 366 hari.' } });
    const localNow=new Date(),today=localNow.toISOString().slice(0,10),yd=new Date(localNow); yd.setUTCDate(yd.getUTCDate()-1); const yesterday=yd.toISOString().slice(0,10),monthStart=`${today.slice(0,7)}-01`,pme=new Date(`${monthStart}T00:00:00.000Z`); pme.setUTCDate(pme.getUTCDate()-1); const previousMonthEnd=pme.toISOString().slice(0,10),previousMonthStart=`${previousMonthEnd.slice(0,7)}-01`;
    const period=(from,to)=>db.get(`SELECT COUNT(*) transactions,COALESCE(SUM(total_amount),0) nominal,COALESCE(SUM(CASE WHEN status='PAID' THEN total_amount ELSE 0 END),0) paid_nominal FROM payments WHERE created_at>=? AND created_at<=?`,[`${from}T00:00:00.000Z`,`${to}T23:59:59.999Z`]);
    const [totals,seriesRows,merchantCount,pToday,pYesterday,pCurrent,pPrevious]=await Promise.all([db.get(`SELECT COUNT(*) total,SUM(CASE WHEN status='PAID' THEN 1 ELSE 0 END) paid,SUM(CASE WHEN status='PENDING' THEN 1 ELSE 0 END) pending,COALESCE(SUM(CASE WHEN status='PAID' THEN total_amount ELSE 0 END),0) volume,COALESCE(SUM(CASE WHEN status='PENDING' THEN total_amount ELSE 0 END),0) clearing FROM payments`),db.all(`SELECT SUBSTRING(created_at,1,10) date,COUNT(*) transactions,COALESCE(SUM(total_amount),0) nominal FROM payments WHERE created_at>=? AND created_at<=? GROUP BY SUBSTRING(created_at,1,10) ORDER BY date`,[`${start.toISOString().slice(0,10)}T00:00:00.000Z`,`${end.toISOString().slice(0,10)}T23:59:59.999Z`]),db.get('SELECT COUNT(*) count FROM merchants WHERE is_active=1'),period(today,today),period(yesterday,yesterday),period(monthStart,today),period(previousMonthStart,previousMonthEnd)]);
    const byDate=new Map(seriesRows.map(row=>[row.date,row])),series=[]; for(let cursor=new Date(start);cursor<=end;cursor.setUTCDate(cursor.getUTCDate()+1)){const date=cursor.toISOString().slice(0,10);series.push(byDate.get(date)||{date,transactions:0,nominal:0});}
    res.json({data:{...totals,available_balance:totals.volume,clearing_balance:totals.clearing,merchants:merchantCount.count,periods:{today:pToday,yesterday:pYesterday,current_month:pCurrent,previous_month:pPrevious},range:{from:start.toISOString().slice(0,10),to:end.toISOString().slice(0,10)},series}});
  } catch(error){next(error);} });
  router.get('/payments/export.csv', auth, requirePermission('payments:read'), async (req, res, next) => { try { const { where, params } = paymentFilters(req.query); const rows = await db.all(`SELECT p.*,m.name merchant_name FROM payments p JOIN merchants m ON m.id=p.merchant_id ${where} ORDER BY p.created_at DESC LIMIT 10000`, params); const headers = ['Payment ID','Order ID','Merchant','Provider Reference','Status','Method','Amount','Total Amount','Customer','Created At','Paid At']; const body = rows.map((p) => [p.id,p.merchant_order_id,p.merchant_name,p.provider_reference,p.status,p.payment_method,p.amount,p.total_amount,p.customer_name,p.created_at,p.paid_at].map(csvCell).join(',')).join('\n'); res.set({ 'Content-Type': 'text/csv; charset=utf-8', 'Content-Disposition': `attachment; filename="evopay-payments-${new Date().toISOString().slice(0,10)}.csv"` }); res.send(`\uFEFF${headers.map(csvCell).join(',')}\n${body}`); } catch(error){next(error);} });
  router.get('/payments', auth, requirePermission('payments:read'), async (req, res, next) => { try { const page = Math.max(1, Number.parseInt(req.query.page, 10) || 1); const limit = Math.min(100, Math.max(10, Number.parseInt(req.query.limit, 10) || 20)); const { where, params } = paymentFilters(req.query); const summary = await db.get(`SELECT COUNT(*) total,COALESCE(SUM(p.total_amount),0) value FROM payments p ${where}`, params); const rows = await db.all(`SELECT p.*,m.name merchant_name FROM payments p JOIN merchants m ON m.id=p.merchant_id ${where} ORDER BY p.created_at DESC LIMIT ? OFFSET ?`, [...params, limit, (page - 1) * limit]); res.json({ data: rows.map(adminPayment), meta: { page, limit, total: Number(summary.total), pages: Math.max(1, Math.ceil(Number(summary.total) / limit)), value: summary.value } }); } catch(error){next(error);} });
  router.get('/payments/:paymentId', auth, requirePermission('payments:read'), async (req, res, next) => { try { const payment = await db.get('SELECT p.*,m.name merchant_name FROM payments p JOIN merchants m ON m.id=p.merchant_id WHERE p.id=?', [req.params.paymentId]); if (!payment) return res.status(404).json({ error: { message: 'Pembayaran tidak ditemukan.' } }); const [events,deliveries,refunds,notes]=await Promise.all([db.all('SELECT id,provider,event_type,signature_valid,payload_json,created_at FROM payment_events WHERE payment_id=? ORDER BY created_at DESC',[payment.id]),db.all('SELECT * FROM webhook_deliveries WHERE payment_id=? ORDER BY created_at DESC',[payment.id]),db.all('SELECT * FROM refunds WHERE payment_id=? ORDER BY created_at DESC',[payment.id]),db.all('SELECT n.*,u.name author_name FROM payment_notes n LEFT JOIN admin_users u ON u.id=n.author_id WHERE n.payment_id=? ORDER BY n.created_at DESC',[payment.id])]); for(const delivery of deliveries) delivery.attempts=await db.all('SELECT * FROM webhook_attempts WHERE delivery_id=? ORDER BY attempt_number DESC',[delivery.id]); res.json({ data: { payment: adminPayment(payment), events, deliveries, refunds, notes } }); } catch(error){next(error);} });
  router.post('/payments/:paymentId/sync', auth, requirePermission('payments:read'), async (req, res, next) => { try { const payment = await db.get('SELECT * FROM payments WHERE id=?',[req.params.paymentId]); if (!payment) return res.status(404).json({ error: { message: 'Pembayaran tidak ditemukan.' } }); if (payment.status !== 'PENDING') return res.status(409).json({ error: { message: 'Hanya pembayaran PENDING yang dapat disinkronkan.' } }); const updated = await paymentService.sync(payment); res.json({ data: publicPayment(updated) }); } catch (error) { next(error); } });
  router.get('/merchants', auth, requirePermission('merchants:read'), async (req, res, next) => { try { const rows=await db.all('SELECT id,name,callback_url,is_active,created_at FROM merchants ORDER BY created_at DESC'); res.json({data:rows.map(safeMerchant)}); } catch(error){next(error);} });
  router.get('/merchants/:merchantId/payment-methods', auth, requirePermission('merchants:read'), async (req,res,next)=>{ try { if(!await db.get('SELECT id FROM merchants WHERE id=?',[req.params.merchantId]))return res.status(404).json({error:{message:'Merchant tidak ditemukan.'}}); res.json({data:await paymentMethods.list(req.params.merchantId)}); } catch(error){next(error);} });
  router.patch('/merchants/:merchantId/payment-methods/:code', auth, requirePermission('merchants:write'), async (req,res,next)=>{ try { if(!await db.get('SELECT id FROM merchants WHERE id=?',[req.params.merchantId]))return res.status(404).json({error:{message:'Merchant tidak ditemukan.'}}); res.json({data:await paymentMethods.update(req.params.merchantId,req.params.code,req.body)}); } catch(error){next(error);} });
  router.get('/merchants/:merchantId/ip-allowlist',auth,requirePermission('merchants:read'),async(req,res,next)=>{try{if(!await merchantById(req.params.merchantId,res))return;res.json(await ipAllowlistService.list(req.params.merchantId));}catch(e){next(e)}});
  router.post('/merchants/:merchantId/ip-allowlist',auth,requirePermission('merchants:write'),async(req,res,next)=>{try{if(!await merchantById(req.params.merchantId,res))return;const data=await ipAllowlistService.create(req.params.merchantId,req.body||{},req.admin.user_id);await auditIpRule(req,'IP_ALLOWLIST_CREATED',data);res.status(201).json({data});}catch(error){next(error);}});
  router.patch('/merchants/:merchantId/ip-allowlist/:ruleId',auth,requirePermission('merchants:write'),async(req,res,next)=>{try{if(!await merchantById(req.params.merchantId,res))return;const data=await ipAllowlistService.toggle(req.params.merchantId,req.params.ruleId,req.body?.is_active);await auditIpRule(req,'IP_ALLOWLIST_UPDATED',data);res.json({data});}catch(error){next(error);}});
  router.delete('/merchants/:merchantId/ip-allowlist/:ruleId',auth,requirePermission('merchants:write'),async(req,res,next)=>{try{if(!await merchantById(req.params.merchantId,res))return;const data=await ipAllowlistService.remove(req.params.merchantId,req.params.ruleId);await auditIpRule(req,'IP_ALLOWLIST_DELETED',data);res.json({data});}catch(error){next(error);}});
  router.post('/merchants', auth, requirePermission('merchants:write'), async (req, res, next) => { try { const name = req.body?.name?.trim(); const callbackUrl = req.body?.callback_url?.trim() || null; if (!name) return res.status(422).json({ error: { message: 'Nama merchant wajib diisi.' } }); if (!validCallbackUrl(callbackUrl)) return res.status(422).json({ error: { message: 'Callback URL harus memakai http:// atau https://.' } }); const apiKey = newApiKey(); const merchant = { id: id('m'), name, callback_url: callbackUrl, is_active: true, created_at: now() }; await db.run('INSERT INTO merchants (id,name,api_key_hash,callback_url,created_at) VALUES (?,?,?,?,?)',[merchant.id,merchant.name,hashApiKey(apiKey),merchant.callback_url,merchant.created_at]); await paymentMethods.seedMerchant(merchant.id); res.status(201).json({data:{...merchant,api_key:apiKey}}); } catch(error){next(error);} });
  router.patch('/merchants/:merchantId', auth, requirePermission('merchants:write'), async(req,res,next)=>{try{const merchant=await merchantById(req.params.merchantId,res);if(!merchant)return;const name=req.body?.name?.trim();const callbackUrl=req.body?.callback_url?.trim()||null;if(!name)return res.status(422).json({error:{message:'Nama merchant wajib diisi.'}});if(!validCallbackUrl(callbackUrl))return res.status(422).json({error:{message:'Callback URL harus memakai http:// atau https://.'}});await db.run('UPDATE merchants SET name=?,callback_url=? WHERE id=?',[name,callbackUrl,merchant.id]);res.json({data:safeMerchant(await db.get('SELECT * FROM merchants WHERE id=?',[merchant.id]))})}catch(e){next(e)}});
  router.delete('/merchants/:merchantId', auth, requirePermission('merchants:write'), async(req,res,next) => { try { const data=await deleteMerchant({merchantId:req.params.merchantId,confirmationName:req.body?.confirmation_name,actorId:req.admin.user_id,ip:req.ip,userAgent:req.get('user-agent')}); res.json({data}); } catch(error) { next(error); } });
  router.post('/merchants/:merchantId/rotate-key', auth, requirePermission('merchants:write'), async(req,res,next)=>{try{const merchant=await merchantById(req.params.merchantId,res);if(!merchant)return;const apiKey=newApiKey();await db.run('UPDATE merchants SET api_key_hash=? WHERE id=?',[hashApiKey(apiKey),merchant.id]);res.json({data:{merchant:safeMerchant(merchant),api_key:apiKey}})}catch(e){next(e)}});
  router.post('/merchants/:merchantId/toggle-status', auth, requirePermission('merchants:write'), async(req,res,next)=>{try{const merchant=await merchantById(req.params.merchantId,res);if(!merchant)return;const isActive=merchant.is_active?0:1;await db.run('UPDATE merchants SET is_active=? WHERE id=?',[isActive,merchant.id]);res.json({data:safeMerchant({...merchant,is_active:isActive})})}catch(e){next(e)}});
  router.post('/deliveries/:deliveryId/retry', auth, requirePermission('webhooks:write'), async (req, res, next) => { try { const item = await db.get('SELECT * FROM webhook_deliveries WHERE id=?',[req.params.deliveryId]); if (!item) return res.status(404).json({ error: { message: 'Callback tidak ditemukan.' } }); await db.run('UPDATE webhook_deliveries SET status=?,last_error=NULL,next_attempt_at=?,updated_at=? WHERE id=?',['PENDING',now(),now(),item.id]); void webhookService.deliver(item.id); res.status(202).json({ data: { queued: true, delivery_id: item.id } }); } catch(error){next(error);} });
  router.get('/webhooks', auth, requirePermission('webhooks:read'), async (req,res,next)=>{ try { const clauses=[];const params=[];for(const [key,column] of [['status','d.status'],['merchant_id','p.merchant_id'],['event_type','d.event_type']])if(req.query[key]){clauses.push(`${column}=?`);params.push(req.query[key])}const where=clauses.length?`WHERE ${clauses.join(' AND ')}`:'';const [data,summary]=await Promise.all([db.all(`SELECT d.*,p.merchant_order_id,p.merchant_id,m.name merchant_name FROM webhook_deliveries d JOIN payments p ON p.id=d.payment_id JOIN merchants m ON m.id=p.merchant_id ${where} ORDER BY d.updated_at DESC LIMIT 500`,params),db.get("SELECT COUNT(*) total,SUM(status='DELIVERED') delivered,SUM(status IN ('RETRYING','DEAD_LETTER')) failed,COALESCE(AVG(attempt_count),0) average_attempts FROM webhook_deliveries")]);res.json({data,summary:{total:Number(summary.total||0),delivered:Number(summary.delivered||0),failed:Number(summary.failed||0),average_attempts:Number(summary.average_attempts||0)}})} catch(error){next(error);} });
  router.get('/webhooks/:deliveryId', auth, requirePermission('webhooks:read'), async (req,res,next)=>{ try { const delivery=await db.get('SELECT d.*,p.merchant_order_id,m.name merchant_name FROM webhook_deliveries d JOIN payments p ON p.id=d.payment_id JOIN merchants m ON m.id=p.merchant_id WHERE d.id=?',[req.params.deliveryId]);if(!delivery)return res.status(404).json({error:{message:'Delivery tidak ditemukan.'}});const attempts=await db.all('SELECT * FROM webhook_attempts WHERE delivery_id=? ORDER BY attempt_number DESC',[delivery.id]);res.json({data:{delivery,attempts}})} catch(error){next(error);} });
  router.get('/search', auth, requirePermission('payments:read'), async(req,res,next)=>{try{const query=req.query.q?.trim();if(!query)return res.json({data:[]});const value=`%${query}%`;const [payments,merchants]=await Promise.all([db.all("SELECT p.id,p.merchant_order_id label,p.status meta,'payment' type FROM payments p WHERE p.id LIKE ? OR p.merchant_order_id LIKE ? OR p.provider_reference LIKE ? LIMIT 8",[value,value,value]),db.all("SELECT id,name label,CASE WHEN is_active=1 THEN 'ACTIVE' ELSE 'DISABLED' END meta,'merchant' type FROM merchants WHERE id LIKE ? OR name LIKE ? LIMIT 5",[value,value])]);res.json({data:[...payments,...merchants]})}catch(e){next(e)}});
  router.post('/payments/:paymentId/notes', auth, requirePermission('payments:note'), async(req,res,next)=>{try{const body=req.body?.body?.trim();if(!body)return res.status(422).json({error:{message:'Catatan wajib diisi.'}});const note={id:id('note'),payment_id:req.params.paymentId,author_id:req.admin.user_id,body,created_at:now()};await db.run('INSERT INTO payment_notes (id,payment_id,author_id,body,created_at) VALUES (?,?,?,?,?)',Object.values(note));res.status(201).json({data:note})}catch(e){next(e)}});
  router.get('/audit-logs', auth, requirePermission('audit:read'), async(req,res,next)=>{try{res.json({data:await db.all('SELECT a.*,u.name actor_name FROM audit_logs a LEFT JOIN admin_users u ON u.id=a.actor_id ORDER BY a.created_at DESC LIMIT 500')})}catch(e){next(e)}});
  router.get('/analytics', auth, requirePermission('analytics:read'), async(req,res,next)=>{try{const [status,methods,hourly,totals]=await Promise.all([db.all('SELECT status,COUNT(*) count,COALESCE(SUM(total_amount),0) nominal FROM payments GROUP BY status'),db.all('SELECT payment_method,COUNT(*) count,COALESCE(SUM(total_amount),0) nominal FROM payments GROUP BY payment_method ORDER BY nominal DESC'),db.all("SELECT SUBSTRING(created_at,12,2) hour,COUNT(*) count FROM payments GROUP BY SUBSTRING(created_at,12,2) ORDER BY hour"),db.get("SELECT COUNT(*) count,COALESCE(AVG(total_amount),0) average,SUM(status='PAID') paid FROM payments")]);res.json({data:{status,methods,hourly,average_transaction_value:totals.average,success_rate:totals.count?totals.paid/totals.count*100:0}})}catch(e){next(e)}});
  return router;
}